requirement.txt
flask
flask-cors
tensorflow==2.7.0
pillow
